+++ src/components/shots/ShotListButton.jsx
@@ -5,12 +5,10 @@ import {
   PopoverProvider,
 } from '@ariakit/react';
 
-import MoreVertIcon from '@material-symbols/svg-400/rounded/more_vert.svg';
+import ListAltIcon from '@material-symbols/svg-400/rounded/list_alt.svg';
 
 import { useSceneNumber } from '@/serdes/UseResolveSceneNumber';
 import PopoverStyle from '@/styles/Popover.module.css';
-
-import SettingsSceneShotsDetailButton from '../scenes/settings/SettingsSceneShotsDetailButton';
 
 /**
  * @param {object} props
@@ -29,17 +27,15 @@ export default function ShotListButton({
   return (
     <div
       className={
-        'flex gap-2 items-center px-2 rounded italic bg-white' + ' ' + className
+        'flex gap-2 items-center mx-4 my-2 px-2 rounded-xl italic bg-gray-100 opacity-60' +
+        ' ' +
+        className
       }>
       <span className="flex-1" />
-      <SettingsSceneShotsDetailButton
-        documentId={documentId}
-        sceneId={sceneId}
-      />
       <PopoverProvider>
         <PopoverDisclosure className="italic flex gap-1">
+          <ListAltIcon className="inline-block w-6 h-6 fill-current" />
           {'Shot List ' + sceneNumber}
-          <MoreVertIcon className="inline-block w-6 h-6 fill-current" />
         </PopoverDisclosure>
         <Popover className={PopoverStyle.popover} modal={true}>
           <PopoverArrow className={PopoverStyle.arrow} />