+++ src/stores/document/value/ShotTypes.js
@@ -1,1 +1,1 @@
-export const SHOT_TYPES = ['', 'WS', 'MS', 'CU', 'FS', 'LS'];
+export const SHOT_TYPES = ['', 'WS', 'MS', 'CU'];